(feat) add DataRequirement test to specifiy a DataType with autorized range of values

+++ test/learn/stickGameWithOpponentDupDouble.cpp
@@ -50,7 +50,7 @@ Evaluation::LearningEnvironment* StickGameWithOpponentD::clone() const
 void StickGameWithOpponentD::doAction(const Data::DataView& action)
 {
     ReinforcementEnvironment::doAction(action);
-    int actionInt = action.getScalar<int>();
+    int actionInt = action.getScalar<size_t>();
 
     // if the game is not over
     if (!this->isTerminal()) {
